cancel ticket validations added

+++ bms/src/main/java/com/nrifintech/bms/controller/TicketController.java
@@ -1,11 +1,13 @@
 package com.nrifintech.bms.controller;
 
+import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.util.Date;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -73,11 +75,27 @@ public class TicketController {
 	}
 	
 	@GetMapping("/cancel/{pnrNo}")
-	public ModelAndView cancelTicket(@PathVariable("pnrNo") String pnrNo) {
+	public ModelAndView cancelTicket(@PathVariable("pnrNo") String pnrNo, HttpServletResponse response,
+			HttpServletRequest request) throws IOException {
 		
-		ticketService.deleteByID(pnrNo);
-		ModelAndView modelAndView = new ModelAndView("redirect:/user/myTickets");
-		return modelAndView;
+		HttpSession session = request.getSession();
+		Object attribute = session.getAttribute("isValidUser");
+		
+		if (attribute != (Object) true) {
+			return new ModelAndView("redirect:/user/login");
+		} else {
+			Ticket ticket = ticketService.getById(pnrNo);
+			int userIdOfTicket = ticket.getUser().getUserid();
+			
+			if(userIdOfTicket == (Integer) session.getAttribute("userid")) {
+				System.out.println("cancel ticket api...");
+				ticketService.deleteByID(pnrNo);
+			}
+			
+			ModelAndView modelAndView = new ModelAndView("redirect:/user/myTickets");
+			return modelAndView;
+		}
+		
 	}
 
 }
